Reestructuracion de los modelos talk
